[backend] refacto(PublicRoadmap): Extract XTM Suite roadmap from service to its own module
