+++ apps/portal-api/src/modules/xtm-suite-roadmap/epic.app.ts
@@ -6,17 +6,20 @@ import {
   QueryEpicsArgs,
   ServiceDefinitionIdentifier,
   UpdateEpicInput,
-} from '../../../__generated__/resolvers-types';
-import portalConfig from '../../../config';
-import { requestContext } from '../../../context/request.context';
-import Epic, { EpicId } from '../../../model/kanel/public/Epic';
-import User from '../../../model/kanel/public/User';
-import { sendMail } from '../../../server/mail-service';
-import { MinIOClient } from '../../../thirdparty/minio/client';
-import { logApp } from '../../../utils/app-logger.util';
-import { processUploads, Upload } from '../document/document.uploads.helper';
-import { DocumentDomain } from '../document/domain/document.domain';
-import { loadSubscribedServiceInstancesByIdentifier } from '../service-instance.domain';
+} from '../../__generated__/resolvers-types';
+import portalConfig from '../../config';
+import { requestContext } from '../../context/request.context';
+import Epic, { EpicId } from '../../model/kanel/public/Epic';
+import User from '../../model/kanel/public/User';
+import { sendMail } from '../../server/mail-service';
+import { MinIOClient } from '../../thirdparty/minio/client';
+import { logApp } from '../../utils/app-logger.util';
+import {
+  processUploads,
+  Upload,
+} from '../services/document/document.uploads.helper';
+import { DocumentDomain } from '../services/document/domain/document.domain';
+import { loadSubscribedServiceInstancesByIdentifier } from '../services/service-instance.domain';
 import { EpicDomain } from './epic.domain';
 
 const addImage = async (user: User, uploads: Upload[]) => {
