+++ apps/portal-api/src/modules/xtm-suite-roadmap/epic.resolver.ts
@@ -1,10 +1,10 @@
 import { toGlobalId } from 'graphql-relay/node/node.js';
-import { Resolvers } from '../../../__generated__/resolvers-types';
-import { DocumentId } from '../../../model/kanel/public/Document';
-import Epic, { EpicId } from '../../../model/kanel/public/Epic';
-import { UnknownErrorCode } from '../../../utils/error/error.code';
-import { mapToGraphQLError } from '../../../utils/error/error.mapping';
-import { DocumentDomain } from '../document/domain/document.domain';
+import { Resolvers } from '../../__generated__/resolvers-types';
+import { DocumentId } from '../../model/kanel/public/Document';
+import Epic, { EpicId } from '../../model/kanel/public/Epic';
+import { UnknownErrorCode } from '../../utils/error/error.code';
+import { mapToGraphQLError } from '../../utils/error/error.mapping';
+import { DocumentDomain } from '../services/document/domain/document.domain';
 import { EpicApp } from './epic.app';
 
 const resolvers: Resolvers = {
