+++ apps/portal-api/src/server/graphql-schema.ts
@@ -15,7 +15,6 @@ import integrationsResolver from '../modules/services/document/opencti/integrati
 import serviceGroupResolver from '../modules/services/group/service-group.resolver';
 import registrationResolver from '../modules/services/registration/registration.resolver';
 import ServiceInstanceResolver from '../modules/services/services.resolver';
-import xtmSuiteRoadmapResolver from '../modules/services/xtm-suite-roadmap/epic.resolver';
 import settingsResolver from '../modules/settings/settings.resolver';
 import useCaseResolver from '../modules/settings/useCase/use-case.resolver';
 import subscriptionsResolver from '../modules/subcription/subscription.resolver';
@@ -23,6 +22,7 @@ import telemetryResolver from '../modules/telemetry/telemetry.resolver';
 import serviceCapabilityResolver from '../modules/user_service/service-capability/service-capability.resolver';
 import userServiceResolver from '../modules/user_service/user_service.resolver';
 import usersResolver from '../modules/users/users.resolver';
+import xtmSuiteRoadmapResolver from '../modules/xtm-suite-roadmap/epic.resolver';
 import nodesResolver from '../nodes/nodes.resolver';
 import { authDirectiveTransformer } from '../security/directive-graphql/directive-auth';
